+++ l2logger/validate.py
@@ -20,17 +20,20 @@ IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 """
 
 import argparse
-import traceback
+import logging
 from pathlib import Path
 
 from tabulate import tabulate
 
 from l2logger import util
 
+logger = logging.getLogger("L2Logger Validator")
+
 
 def run():
     # Instantiate parser
-    parser = argparse.ArgumentParser(description='Validate log format from the command line')
+    parser = argparse.ArgumentParser(
+        description='Validate log format from the command line')
 
     # Log directories can be absolute paths, relative paths, or paths found in $L2DATA/logs
     parser.add_argument('log_dir', type=str, help='Log directory of scenario')
@@ -52,24 +55,32 @@ def run():
     util.validate_log(log_data, logger_info['metrics_columns'])
     print('\nLog format validation passed!\n')
 
+    # Filter data by completed experiences
+    log_data = log_data[log_data['exp_status'] == 'complete']
+
     # Fill in regime number and sort
     log_data = util.fill_regime_num(log_data)
-    log_data = log_data.sort_values(by=['regime_num', 'exp_num']).set_index("regime_num", drop=False)
+    log_data = log_data.sort_values(
+        by=['regime_num', 'exp_num']).set_index("regime_num", drop=False)
 
     # Print log summary
     log_summary = util.parse_blocks(log_data)
     if log_summary['task_params'].dropna().size:
-        log_summary['task_params'] = log_summary['task_params'].apply(lambda x: x[:25] + '...' if len(x) > 25 else x)
+        log_summary['task_params'] = log_summary['task_params'].apply(
+            lambda x: x[:25] + '...' if len(x) > 25 else x)
     else:
         log_summary = log_summary.dropna(axis=1)
 
     print('Log summary:')
-    print(tabulate(log_summary.drop(columns='regime_num'), headers='keys', tablefmt='psql'))
+    print(tabulate(log_summary.set_index("regime_num",
+          drop=True), headers='keys', tablefmt='psql'))
 
 
 if __name__ == '__main__':
+    # Configure logger
+    logging.basicConfig(level=logging.INFO)
+
     try:
         run()
-    except Exception as e:
-        print(f'Error with validating logs: {e}')
-        traceback.print_exc()
+    except (RuntimeError, KeyError) as e:
+        logger.exception(f'Error with validating logs: {e}')